+++ keras/text_classification/transformers/__init__.py
@@ -1,0 +1,7 @@
+from transformers.gensim_word2vec import GensimWord2VecVectorizer
+from transformers.keras_word2vec import KerasWord2VecVectorizer
+
+__all__ = [
+    'GensimWord2VecVectorizer',
+    'KerasWord2VecVectorizer'
+]